\name{r2lh}
\alias{r2lh}
\alias{r2lh-package}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{~ Overview: R to LaTeX / HTML ~}
\description{
  \tabular{ll}{
    Package: \tab r2lh\cr
    Type: \tab Package\cr
    Version: \tab 0.6.1\cr
    Date: \tab 2009-06-01\cr
    License: \tab  GPL (>= 2) \cr
  }
}
\details{
  \code{r2lBiv} performs some bivariate analyses, then generates
  code to be included in a LaTeX document in order to print out the analyses in a (so
  nice!) LaTeX way.
  See \code{vignette(r2lh-exampleOfOutput)} for display details.
}

\section{r2lBiv, in real context...}{
  Bob is working VERY hardly on World of Warcraft... John, his boss, gets in his office:
  \itemize{
    \item Bob, I have some urgent work for you
    \item ...
    \item Bob, I'm talking to you
    \item Huuu ?
    \item I said, I have some urgent work.
    \item Well, John, I am quite busy at the moment.
    \item It is our major client.
    \item Again ?
    \item Yep. He's very happy with the univariate analysis you gave him
      last week. He's now asking for the bivariate analysis.
    \item Hey, I am still a senior statistician, give that to a rookie!
    \item He is still our major client, and now, he loves you !
    \item I need four weeks.
    \item I give you one.
    \item I said four so you are supposed to give me two !
    \item You ask four, that means you need two, I give you one...
  }
  Once more, Bob pouts, internally smiling, and starts again kicking
  some orc ass.
  He knows that running \code{\link{r2lb}} will take him 2 or 3 minutes... As long as John does not know about
  \code{r2lb}, life is beautiful.
}

\section{Author}{
  Christophe Genolini \cr
  \email{christophe.genolini@free.fr}\cr
  INSERM U669 / Maison de Solenn / Paris \cr
  Modal'X / University of Paris Ouest - Nanterre \cr \cr

  Bernard Desgraupes \cr
  \email{bernard.desgraupes@u-paris10.fr}\cr
  Modal'X / University of Paris Ouest - Nanterre \cr \cr
}

\references{LaTeX web site \url{http://www.latex-project.org/}\cr
Data are available on line: \url{http://christophe.genolini.free.fr/EPO/EPO2007-Fraude.php}}


\seealso{
  \code{\link{r2latexUniv-package}},
  \code{\link{r2latexUniv}},
  \code{\link{r2latexBiv}},
  \code{\link{r2htmlUniv}},
  \code{\link{r2htmlBiv}},
  \code{\link{r2lMainFile}},
  \code{\link{r2hMainFile}},
  \code{\link{examCheating}},
  \code{\link{Sweave}},
  \code{\link[Hmisc]{latex}}
}

\keyword{print}        % Printing
\keyword{package}      % Package summary
\keyword{classes}      % Data type
\keyword{programming}  % Programming
\keyword{interface}    % interface to other languages
\keyword{utilities}    % Utilities
\keyword{univar}       % Simple univariate statistics




