\name{r2lUniv}
\alias{r2latexUniv-package}
\alias{r2htmlUniv-package}

%- Also NEED an '\alias' for EACH other topic documented here.
\title{~ Overview: R to LaTeX/HTML, Univariate Analyses ~}
\description{
  \tabular{ll}{
    Package: \tab r2lUniv\cr
    Type: \tab Package\cr
    Version: \tab 0.9.4\cr
    Date: \tab 2009-1-12\cr
    License: \tab  GPL (>= 2) \cr
  }
}
\details{
  \code{r2lUniv} performs some basic analyses, then generates a
  code to be included in a LaTeX document to print the analyses in a (so
  nice!) LaTeX way. See \url{/library/r2lh/doc/r2lhOutput.pdf} for
  display details.

}

\section{r2lUniv, in real context...}{
  Bob is working hardly on his Tetris (or on Minesweeper,
  depends). John, his boss, gets in his office:
  \itemize{
    \item Bob, I have some urgent work for you
    \item ...
    \item Bob, I'm talking to you
    \item Huuu ?
    \item I said, I have some urgent work.
    \item Well, John, I am quite busy at the moment.
    \item It is our major client. He's asking for the univariate analyses of his data.
    \item Hey, I am a senior statistician, give that to a rookie!
    \item It is our major client, I want something fast and perfect. You're
    the only one that produces LaTeX reports.
    \item Hmmm. How many variables?
    \item 485.
    \item I need two weeks.
    \item I give you one. I want your report by Friday 11th , 13:00
  }
  Bob pouts, internally smiling, and starts his Tetris again.
  He knows that running \code{\link{r2lu}} on Friday 11th arround 12:30 will
  be soon enough... As long as John does not know about
  \code{r2lu}, life is beautiful.

  See \url{/library/r2lh/doc/r2lhOutput.pdf} for display details.
}

\section{Author}{
  Christophe Genolini \cr
  \email{christophe.genolini@free.fr}\cr
  PSIGIAM: Paris Sud Innovation Group in Adolescent Mental Health \cr
  INSERM U669 / Maison de Solenn / Paris\cr
  \cr
  Bernard Desgraupes \cr
  \email{bernard.desgraupes@u-paris10.fr}\cr
  MODAL'X / University of Paris Ouest - Nanterre
}

\references{LaTeX web site \url{http://www.latex-project.org/}\cr
Data are available on line: \url{http://christophe.genolini.free.fr/EPO/EPO2007-Fraude.php}}


\seealso{
  \code{\link{r2latexUniv}},
  \code{\link{r2latexBiv}},
  \code{\link{r2htmlUniv}},
  \code{\link{r2htmlBiv}},
  \code{\link{r2lMainFile}},
  \code{\link{r2hMainFile}},
  \code{\link{examCheating}},
  \code{\link{r2lh}},
  \code{\link{Sweave}},
  \code{\link[Hmisc]{latex}}
}

\keyword{print}        % Printing
\keyword{package}      % Package summary
\keyword{classes}      % Data type
\keyword{programming}  % Programming
\keyword{interface}    % interface to other languages
\keyword{utilities}    % Utilities
\keyword{univar}       % Simple univariate statistics




